dotenv with no trailing newline fix
